Added Export option, Merged same date Events Allowed Decimal in amount

+++ app/src/main/java/athato/ghummakd/jigayasa/domain/model/Event.kt
@@ -6,6 +6,6 @@ data class Event(
     val message: String,
     val timestamp: Long,
     val category: String? = null,
-    val amount: Long? = null,
+    val amount: Double? = null,
     val currencyCode: String? = null
 )
